requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
mysqlclient==2.2.0
Pillow==10.1.0
django-friendship==1.9.6
